[Refactor] #32 다이얼로그 수정
